db_control/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
regex
bcrypt